ptrdiff_t is a C99 stdint.h type - substitute ssize correctly
The posix ssize_t has never been standard C but we need to consume nghttp2 and libevent
which make heavy use of these sizes in public headers. Use this type on windows over
other Windows-specific types such as SSIZE_T.

+++ include/envoy/common/platform.h
@@ -9,7 +9,7 @@
   __pragma(pack(pop))
 
 #ifdef _M_X64
-using ssize_t = int64_t;
+using ssize_t = ptrdiff_t;
 #else
 #error Envoy is not supported on 32-bit Windows
 #endif
